Refactor data loading function
