+++ excel_analysis/stock_processing.py
@@ -18,7 +18,7 @@ from excel_analysis.constants import (
     INDEX_COLUMN,
     TRAIN_TEST_SPLIT_RATIO,
 )
-from excel_analysis.utils.data_loaders import get_valid_sheets, load_data
+from excel_analysis.utils.data_loaders import validate_and_load_sheets
 from excel_analysis.models.neural_networks import obtener_threshold_optimo
 from excel_analysis.utils.grading_system import (
     assign_stock_grade,
@@ -95,25 +95,6 @@ def process_stock_data(df, sheet_name, results_list, columns):
     logging.info(
         f"💰 Valor final de esta hoja: {final_value}, Threshold: {optimal_threshold}, Grado: {stock_grade}"
     )
-
-
-def validate_and_load_sheets(file_name, index_column):
-    valid_sheets = get_valid_sheets(file_name, index_column)
-    if not valid_sheets:
-        logging.error("No se encontraron hojas válidas en el archivo Excel.")
-        return None, None
-
-    logging.info(
-        f"📂 Encontramos {len(valid_sheets)} hojas válidas en el archivo Excel\n"
-    )
-
-    all_data = load_data(
-        file_name=file_name,
-        index_column=index_column,
-        sheets_to_load=valid_sheets,
-        single_sheet=False,
-    )
-    return valid_sheets, all_data
 
 
 def assign_grades_and_update_results(results):
